refactor: use cobra
