+++ go.mod
@@ -3,8 +3,7 @@ module github.com/cosmos/cosmos-proto
 go 1.16
 
 require (
-	github.com/kr/pretty v0.1.0 // indirect
+	github.com/spf13/cobra v1.2.1
 	github.com/stretchr/testify v1.7.0
 	google.golang.org/protobuf v1.27.0
-	gopkg.in/check.v1 v1.0.0-20180628173108-788fd7840127 // indirect
 )
